(1) Extract PanelGrid as reusable component (2) Make RightSideOverlay fit on the right side of the screen
